Refactor AddCommand to AddStockCommand
Also moved class from 'commands' package to 'add' sub-package.
